requirements.txt
pip~=24.0
Django~=5.0.4
asgiref~=3.8.1
sqlparse~=0.5.0
colorama~=0.4.6
requests~=2.31.0
djangorestframework~=3.15.2
python-dotenv~=1.0.1
django-filter~=24.2